app/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
debugpy==1.8.17
defusedxml==0.7.1
Django==5.2.5
django-cors-headers==4.9.0
django-csp==4.0
django-db-schema-renderer==0.1.2
django-extensions==3.2.1
django-filter==25.2
django-health-check==3.20.0
django-permissions-policy==4.28.0
django-spaghetti-and-meatballs==0.4.2
djangorestframework==3.16.1
djangorestframework-csv==3.0.2
djangorestframework-gis==1.2.0
djangorestframework-xml==2.0.0
drf-extensions==0.8.0
drf-writable-nested==0.7.2
drf-spectacular==0.28.0
drf-extra-fields==3.7.0
drf_base64
factory-boy==3.2.1
filetype==1.2.0
idna==3.7
model-bakery==1.20.5
nh3==0.2.18
Pillow==9.4.0
pillow-heif==0.13.0
psycopg2>=2.9.5
graphviz==0.20.1
pygraphviz==1.11
pytz==2022.7.1
requests==2.32.5
six==1.16.0
sqlparse==0.5.0
unicodecsv==0.14.1
urllib3==2.5.0
uWSGI==2.0.25.1
djangorestframework-simplejwt==5.5.1
python-box[all]~=7.0
sorl-thumbnail==12.8.0
django-rename-app==0.1.6
celery==5.5.3
django-celery-beat==2.8.1
django-celery-results==2.6.0
redis==4.5.5
django_redis==5.4.0
python-magic==0.4.27
cryptography==44.0.1
mozilla-django-oidc==3.0.0
PyJWT<2.4.0
django-session-timeout-joinup==1.0.0

# development
django-debug-toolbar==6.0.0
Faker==8.13.2

# tests
requests_mock==1.12.1

# monitoring
django-prometheus==2.3.1
prometheus-client==0.20.0
shapely==2.0.4
git+https://github.com/forza-mor-rotterdam/mor-api-services.git@58ab7c950b4751acb34b9208529c85f0d9c46c05
pika==1.3.2
